+++ contracts/IWallet.sol
@@ -18,5 +18,5 @@ interface IWallet {
      *      can be withdrawn anytime using "entryPoint.withdrawTo()"
      *      In case there is a paymaster in the request (or the current deposit is high enough), this value will be zero.
      */
-    function validateUserOp(UserOperation calldata userOp, bytes32 requestId, uint requiredPrefund) external;
+    function validateUserOp(UserOperation calldata userOp, bytes32 requestId, uint256 requiredPrefund) external;
 }
